% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B.r
\name{B}
\alias{B}
\title{B family functions extracts the specified substrings in the given index ranges based on numbers or characters.
If a character is given, the function finds its location and treats its origin as a boundary.
B is vectorised version of B0}
\usage{
B(XS, X_START, X_END)
}
\description{
B family functions extracts the specified substrings in the given index ranges based on numbers or characters.
If a character is given, the function finds its location and treats its origin as a boundary.
B is vectorised version of B0
}
