\name{cargar}
\alias{cargar}
\title{Cargar Archivos de datos}
\arguments{
  \item{header}{Si el archivo tiene nombre de columnas, usar header=T}
  \item{dec}{Si el separador decimal es la coma (","), usar dec=","}
  \item{sep}{Separador de columnas, por defecto sep=";"}
}
\usage{
x<-cargar()
}
\description{
Permite cargar datos sólo seleccionando el archivo con la extensión apropiada (.csv,.txt,.xls,.xlsx)
}
\examples{
x<-cargar()
}
