% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lc50.r
\name{ic50}
\alias{ic50}
\title{Calculates the ic50 values}
\usage{
ic50(
  df,
  abs_col,
  conc_col,
  substance_name_col,
  negative_identifier,
  positive_identifier
)
}
\arguments{
\item{df}{a data.frame which contains all the data}

\item{abs_col}{the name of the column in df which contains the dependent variable}

\item{conc_col}{the name of the column in df which contains the different concentrations}

\item{substance_name_col}{the name of the column in df which contains the different names of the compounds}

\item{negative_identifier}{a character defining the name to identify the negative control within conc_col}

\item{positive_identifier}{a character defining the name to identify the positive control within conc_col}
}
\value{
a list is returned containing the ic50 value the fitted plots and other parameters
}
\description{
Calculates the ic50 values
}
\examples{
path <- system.file("data", package = "MTT")
df <- read.csv(paste0(path, "/ExampleData.txt"))
ic50(df, "abs", "conc", "names", "neg", "pos")
}
