% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S.R
\name{S}
\alias{S}
\title{S family functions search for matches to argument pattern.}
\usage{
S(X, ..., ignore.case = T, perl = T, fixed = F, useBytes = F)
}
\description{
B is vectorised version of B0
First argument - X, a character vector where matches are sought, or an object which can be coerced by as.character to a character vector.
Second arguments is elipsis.
S allow to search by all given pattern.
There is option to exclude pattern by using operator:\code{!} or \code{-}
For example:
XV <- names(sort(islands,decreasing = T))
S(XV,-'A','b') #matching B and not A.
S(XV,-'^A') #matching not started by A.
Default option is ignoring cases.
S return logical values
SV return values
SW return which values
}
